The inventory AddItem and MoveItem methods have been updated. The RoguelikeObject MakeRoguelikeObject function has been updated along with a new MakeRoguelikeObjectTemporary function. Minor updates to the scene.
Overview of the Inventory method changes:
AddItem - Now, instead of taking a freshly created item and distributing it into the inventory, it now takes a reference to an existing stack. It also now takes an int amount which represents how much of the reference stack we want to add to the new inventory. The existing stack StackSize and the amount parameter are decremented as it is added to existing stacks in the new inventory and when a free space is the target of the AddItem, if the amount to move parameter is less than the full stack amount, we create a new stack and add it, If the the amount to move is >= to the full stack amount, we move the whole stack over to the free space.

MoveItem is now pretty much just calling AddItem.

Now when MakeRoguelikeObject is called, the new object is first added to a temporary inventory that should ALWAYS be free. Then we try to move the object to the destination inventory.